Query by position was added using Postgresql + Q3C. the square query methods have been separated into the specific classes of Oracle and Postgresql. Closed #1231
