+++ api/lib/db_postgresql.py
@@ -1,4 +1,7 @@
+from sqlalchemy import Column
 from sqlalchemy.dialects import postgresql
+from sqlalchemy.sql import and_, or_, text
+from sqlalchemy.sql.expression import between, literal_column
 
 
 class DBPostgresql:
@@ -72,3 +75,26 @@ class DBPostgresql:
 
     def get_schema_name(self, schema):
         return schema
+
+    def get_condition_square(self, lowerleft, upperright, property_ra, property_dec):
+
+        raul = float(lowerleft[0])
+        decul = float(upperright[1])
+        ul = "{%s, %s}" % (raul, decul)
+
+        raur = float(upperright[0])
+        decur = float(upperright[1])
+        ur = "{%s, %s}" % (raur, decur)
+
+        ralr = float(upperright[0])
+        declr = float(lowerleft[1])
+        lr = "{%s, %s}" % (ralr, declr)
+
+        rall = float(lowerleft[0])
+        decll = float(lowerleft[1])
+        ll = "{%s, %s}" % (rall, decll)
+
+        # ul, ur, lr, ll
+        stm = "q3c_poly_query(ra, dec, '{ %s, %s, %s, %s}')" % (ul, ur, lr, ll)
+
+        return and_(text(stm)).self_group()
